# Some small improvements to inmem storage