--- a/v1/storage/path.go
+++ b/v1/storage/path.go
@@ -8,40 +8,40 @@ import (
 	"errors"
 	"fmt"
 	"net/url"
+	"slices"
 	"strconv"
 	"strings"
 
 	"github.com/open-policy-agent/opa/v1/ast"
 )
 
+// RootPath refers to the root document in storage.
+var RootPath = Path{}
+
 // Path refers to a document in storage.
 type Path []string
 
 // ParsePath returns a new path for the given str.
 func ParsePath(str string) (path Path, ok bool) {
-	if len(str) == 0 {
-		return nil, false
-	}
-	if str[0] != '/' {
+	if len(str) == 0 || str[0] != '/' {
 		return nil, false
 	}
 	if len(str) == 1 {
 		return Path{}, true
 	}
-	parts := strings.Split(str[1:], "/")
-	return parts, true
+
+	return strings.Split(str[1:], "/"), true
 }
 
 // ParsePathEscaped returns a new path for the given escaped str.
 func ParsePathEscaped(str string) (path Path, ok bool) {
-	path, ok = ParsePath(str)
-	if !ok {
-		return
-	}
-	for i := range path {
-		segment, err := url.PathUnescape(path[i])
-		if err == nil {
-			path[i] = segment
+	if path, ok = ParsePath(str); ok {
+		for i := range path {
+			if segment, err := url.PathUnescape(path[i]); err == nil {
+				path[i] = segment
+			} else {
+				return nil, false
+			}
 		}
 	}
 	return
@@ -49,7 +49,6 @@ func ParsePathEscaped(str string) (path Path, ok bool) {
 
 // NewPathForRef returns a new path for the given ref.
 func NewPathForRef(ref ast.Ref) (path Path, err error) {
-
 	if len(ref) == 0 {
 		return nil, errors.New("empty reference (indicates error in caller)")
 	}
@@ -85,36 +84,17 @@ func NewPathForRef(ref ast.Ref) (path Path, err error) {
 // is less than other, 0 if p is equal to other, or 1 if p is greater than
 // other.
 func (p Path) Compare(other Path) (cmp int) {
-	for i := range min(len(p), len(other)) {
-		if cmp := strings.Compare(p[i], other[i]); cmp != 0 {
-			return cmp
-		}
-	}
-	if len(p) < len(other) {
-		return -1
-	}
-	if len(p) == len(other) {
-		return 0
-	}
-	return 1
+	return slices.Compare(p, other)
 }
 
 // Equal returns true if p is the same as other.
 func (p Path) Equal(other Path) bool {
-	return p.Compare(other) == 0
+	return slices.Equal(p, other)
 }
 
 // HasPrefix returns true if p starts with other.
 func (p Path) HasPrefix(other Path) bool {
-	if len(other) > len(p) {
-		return false
-	}
-	for i := range other {
-		if p[i] != other[i] {
-			return false
-		}
-	}
-	return true
+	return len(other) <= len(p) && p[:len(other)].Equal(other)
 }
 
 // Ref returns a ref that represents p rooted at head.