--- a/v1/test/authz/authz_bench_test.go
+++ b/v1/test/authz/authz_bench_test.go
@@ -61,7 +61,7 @@ func runAuthzBenchmark(b *testing.B, mode InputMode, numPaths int, extras ...boo
 			b.Fatal(err)
 		}
 
-		if err = storage.WriteOne(ctx, store, storage.AddOp, storage.Path{}, data); err != nil {
+		if err = storage.WriteOne(ctx, store, storage.AddOp, storage.RootPath, data); err != nil {
 			b.Fatal(err)
 		}
 	} else {