… v1/storage/path_test.go